Create nginx-ssl.conf
This configuration should be used for Nginx if SSL is desired. It should replace nginx.conf in /usr/local/etc/nginx/ and named nginx.conf based on current conventions. Ideally post_install.sh would be modified to pass an option to enable SSL which would copy this file instead of the non-SSL version. Nice to haves would be specifying crt and key names. I was unable to get "certificate" and "keyfile" in MyConfig.conf to point to the crt and key by a different name.